Smoke test python module

+++ _packages/python/build.py
@@ -11,10 +11,8 @@ def get_repo_height() -> int:
 
 
 def build(kwargs):
-    kwargs['version'] = f"1.0.{get_repo_height()}"
-    print(kwargs)
-
-    output_data = pathlib.Path(os.path.dirname(__file__)).joinpath('./share')
     input_data = pathlib.Path(os.path.dirname(__file__)).joinpath('../../_site/.data')
-    for file in input_data.rglob('*.json'):
-        shutil.copy(file, output_data)
+    if input_data.exists():
+        output_data = pathlib.Path(os.path.dirname(__file__)).joinpath('./share')
+        for file in input_data.rglob('*.json'):
+            shutil.copy(file, output_data)
